Changed float to double, short to long int, int to long int types throughout in GEMIPM2K and in the respective parts of GEMS-PSI (with DODs for MULTI data objects); replaced outdated calls (like memset(), memcpy() etc.) with loops, all this to improve compatibility with parallel architectures (SD)
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@1122 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/s_fgl.h
@@ -95,7 +95,7 @@ class EOSPARAM
   double M2R(unsigned long int i) {return m2par[i];};
   double A(unsigned long int i)  {return apar[i];};
 
-  unsigned ParamMix( double *Xin);
+  unsigned long int ParamMix( double *Xin);
   void PureParam(double* ,double*,double*,double* );
   long int NCmp(){return NComp;};
 };
@@ -307,8 +307,8 @@ protected:
      double &fugmix, double &zmix, double &vmix);
 	long int FugacitySpec( double *fugpure, double *params  );
 
-	int GetEosParam( float *params ); // Loads EoS parameters for NComp species
-	int GetMoleFract( double *Wx ); // Loads mole fractions for NComp species
+	long int GetEosParam( float *params ); // Loads EoS parameters for NComp species
+	long int GetMoleFract( double *Wx ); // Loads mole fractions for NComp species
 	double ObtainResults( double *ActCoef ); // returns activity coeffs and phase volume
 
 };
